perf: reduce call count and data quantity for high use query
